requirements.txt
llama-index-embeddings-gemini
llama-index-llms-gemini
llama-index
google-generativeai
streamlit
